requirements.txt
cffi==2.0.0
cryptography==46.0.3
pycparser==2.23
pyOpenSSL==25.3.0
typing_extensions==4.15.0
